refactor(frontend): app components
Move shared app components into the `app/` directory.
