+++ frontend/src/app/Notice.tsx
