+++ frontend/src/app/asset-list/ExpandableRow.tsx
